added motd + did click magic so the package and the cli can both be called fig + forgor to put it in the branch name but i added the logo
