+++ MANIFEST.in
@@ -1,0 +1,3 @@
+include fig/impact.ttf
+include fig/scripts/logo_ascii_art.txt
+include fig/scripts/fig_ascii_font.txt
