+++ fig/scripts/fig_ascii_font.txt
@@ -1,0 +1,10 @@
+
+
+███████╗██╗ ██████╗
+██╔════╝██║██╔════╝
+█████╗  ██║██║  ███╗
+██╔══╝  ██║██║   ██║
+██║     ██║╚██████╔╝
+╚═╝     ╚═╝ ╚═════╝
+
+
